Convert door Animator to use Synchronisevars
