--- UnityProject/Assets/Scripts/Messages/Client/NewPlayer/DoorNewPlayer.cs
@@ -1,36 +1,0 @@
-﻿using Doors;
-using Mirror;
-
-namespace Messages.Client.NewPlayer
-{
-	public class DoorNewPlayer : ClientMessage<DoorNewPlayer.NetMessage>
-	{
-		public struct NetMessage : NetworkMessage
-		{
-			public uint Door;
-		}
-
-		public override void Process(NetMessage msg)
-		{
-			// LoadNetworkObject returns bool, so it can be used to check if object is loaded correctly
-			if (LoadNetworkObject(msg.Door))
-			{
-				if (NetworkObject.TryGetComponent(out DoorMasterController doorMasterController))
-				{
-					doorMasterController.UpdateNewPlayer(SentByPlayer.Connection);
-				}
-			}
-		}
-
-		public static NetMessage Send(uint netId)
-		{
-			NetMessage msg = new NetMessage
-			{
-				Door = netId
-			};
-
-			Send(msg);
-			return msg;
-		}
-	}
-}
