--- UnityProject/Assets/Scripts/Messages/Server/DoorUpdateMessage.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a3ba2e089c26436982c6fa909eeffc54
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
