+++ UnityProject/Assets/Scripts/Objects/Doors/DoorMasterController.cs
@@ -164,23 +164,7 @@ namespace Doors
 			HackingProcessBase.RegisterPort(TryClose, this.GetType());
 			HackingProcessBase.RegisterPort(ConfirmAIConnection, this.GetType());
 		}
-
-		public override void OnStartClient()
-		{
-			DoorNewPlayer.Send(netId);
-		}
-
-		/// <summary>
-		/// Used when player is joining, tells player to open the door if it is opened.
-		/// </summary>
-		public void UpdateNewPlayer(NetworkConnection playerConn)
-		{
-			DoorUpdateMessage.Send(playerConn, gameObject,
-				IsClosed ? DoorUpdateType.Close : DoorUpdateType.Open,
-				true,
-				ConstructibleDoor != null && ConstructibleDoor.Panelopen);
-		}
-
+		
 
 		private bool CheckPower()
 		{
@@ -474,8 +458,10 @@ namespace Doors
 				return;
 			}
 
+			doorAnimator.PanelOpen = ConstructibleDoor != null && ConstructibleDoor.Panelopen;
 
-			DoorUpdateMessage.SendToAll(gameObject, DoorUpdateType.Close, ConstructibleDoor != null && ConstructibleDoor.Panelopen);
+			doorAnimator.SyncDoorStatus(doorAnimator.SyncDoorUpdateType,DoorAnimatorV2.DoorUpdateType.Close );
+
 
 			if (damageOnClose)
 			{
@@ -510,7 +496,8 @@ namespace Doors
 
 			if (!isPerformingAction)
 			{
-				DoorUpdateMessage.SendToAll(gameObject, DoorUpdateType.Open, ConstructibleDoor != null && ConstructibleDoor.Panelopen);
+				doorAnimator.PanelOpen = ConstructibleDoor != null && ConstructibleDoor.Panelopen;
+				doorAnimator.SyncDoorStatus(doorAnimator.SyncDoorUpdateType,DoorAnimatorV2.DoorUpdateType.Open);
 			}
 
 		}
